move to etl.py
